Add random scattering & fix `psi_index` axes
Fix `psi_index` so that the order of its axes are predictable. Before, an error in the axis order could lead to an error that stops the code from running (but does not lead to any output of erroneous numbers).

Also add `random_scatter` to scatter random pairs of particles a given number of times and return the evolution of the state.

Move the core scattering dynamics to dynamics.py.