Add approach 2
Approach 2 scatters neutrinos as (usually independent) density matrices. Filenames and folder structure are updated a bit to reflect the existence of 2 approaches. This version of the code was used for the 23/10/23 meeting.

The option to scatter neutrinos with random relative angles (in their momenta) is added.

Graphs from the analysis of equilibration time in approach 1 are included.

`random_scatter` is moved from analysis.py to dynamics.py (for organization).

cooking_simulation_1.ipynb is renamed to approach_1_fixed_angles.ipynb.