Fix approach 2 normalization bug
Due to what appears to be floating point error (since the trace of density matrices after scatters are equal to 1 up to the last digit of the floating point number), the trace of density matrices and overall results used to blow up after ~50 scatters. This adds a step to divide the trace out, thus renormalizing the density matrix, after each scattering. It appears to fix the bug.