Add eq. 19 check & variable neutrino fractions
Add notebooks/2024_1_4_eq_19_check.ipynb for verifying the analytic solutions to eq. 19 in the current OILE paper draft, with and without the "extra" 4th term.

Add the ability to adjust the fractions of e and x neutrinos. For example, for 100 total neutrinos, a mix such as 60/40 is now possible, in addition to the 50/50 that previously was the only option.

Add some minor useful things to notebooks/approach_2_multivacum_2.ipynb.