Add simultaneous vacuum oscillations (unfinished)
CAUTION: The simultaneous vacuum oscillations (in `dynamics.scatter`) are performed as if the states are in the mass basis, while the free vacuum oscillations (in `dynamics.propagate`) are still applied according to the flavor basis (as they were before). For this reason, the two shouldn't be used in the same simulation, yet. The plan for the next step is to update everything to be in the mass basis.

Modify the time evolution operator in `dynamics.scatter` to reflect the Hamiltonian that is the sum of the vacuum and interaction Hamiltonians, H = V + Omega. The "free" vacuum oscillations that take place in their own simulation step remain in place as well. The strength of the two types of vacuum oscillations can be adjusted separately, via the variables:
 - `alpha1` = omega * Delta z (for simultaneous vacuum oscillations)
 - `alpha2`= omega * t (for free vacuum oscillations, where t is the average free time of a neutrino between its collisions)
(Each alpha can have different values for the two different flavors of neutrinos.)